LayoutTests/fast/forms/slider-mouse-events-expected.txt

This tests that slider controls get mousedown, mousemove, input, and mouseup events.

mousemove
mousedown
mousemove
input
mouseup
